+++ flutter/CHANGELOG.md
@@ -6,6 +6,19 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## master
+
+### Added
+
+- `MediaDeviceDetails.sampleRate()`, `MediaDeviceDetails.numChannels()` function (only works on Windows right now).
+
+### Fixed
+
+- `MediaDeviceDetails.groupId()` not working on Windows platform.
+
+
+
+
 ## [0.13.1] · 2026-02-06
 [0.13.1]: https://github.com/instrumentisto/medea-jason/tree/medea-jason-0.13.1/flutter